src/components: move DrawerHeader
